Makefile: always define MOFILES, but respect NO_GETTEXT=Y
Fix up a rule changed in 4348824059 (artifacts-tar: respect
NO_GETTEXT, 2021-07-04) to more narrowly fix the problem it was trying
to solve, namely that we shouldn't be bundling up the MOFILES with the
artifacts-tar under NO_GETTEXT=Y.

I suggested doing this in [1], the result of that discussion was (I
think) that there was particular technical reason not to go with this,
but there was the existing patch that become 4348824059.

I now need to modify ARTIFACTS_TAR for some other changes, including
selectively adding things to it. So let's fix this pattern first so we
can make those subsequent changes consistent with this.

1. https://lore.kernel.org/git/878s2m8ns9.fsf@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>