protocol: add a packet_client_error() API
Since the protocol change detailed in 2d103c31c2c (pack-protocol.txt:
accept error packets in any context, 2018-12-29) we can emit "ERR"
packets in any context that requires a response. Add an API for
emitting one, and following that up with a die().

Despite the protocol change in 2d103c31c2c we have not made wide or
consistent use of the "ERR" packet in practice. So e.g. if a client
supplies unknown arguments under HTTP the client will report a cryptic
error about not being able to read the "acknowledgments" section
header.

This is because the remote side simply died, and we are not connected
to the server's STDERR.

Even when we are connected to STDERR (e.g. under ssh://, or file://)
our output is some combination of racy, inconsistent or redundant, as
subsequent fixes and tests will demonstrate.

If we were being paranoid here we would only emit these ERR packets to
clients that understand the spec change detailed in 2d103c31c2c. I'm
not doing that, because:

 1. In practice we supported this for a long time, it just wasn't
    specified. See a8073289793 (connect.c: add a way for git-daemon to
    pass an error back to client, 2008-11-01) and
    d78e5aecf93 (pack-protocol: document "ERR" line, 2011-10-04). I.e.
    Even when we start using this to emit errors to v0/v1 clients they'll
    grok it.

 2. If they don't they're not much worse off, we'll still emit a die()
    message for now, and if and when we stop doing that it's only
    going to be ancient clients that are impacted. In practice they
    probably don't support most of the feature likely to emit these
    errors in practice, e.g. filters.

This is the first time since d162b25f956 (tests: remove support for
GIT_TEST_GETTEXT_POISON, 2021-01-20) that I've found myself slightly
missing GIT_TEST_GETTEXT_POISON. As noted in the packet_client_error()
comments we've got a tricky case here where we'd like to translate
messages to the user, but not the messages on the wire. We'd ideally
have some way to specify the language over the wire, but for now let's
use LANG=C there.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>