protocol v2 tests: test exact stdout/stderr for .gitmodules fsck
Change the test added in 5476e1efded (fetch-pack: print and use
dangling .gitmodules, 2021-02-22) to demonstrate that we redundantly
print the same warning added in that recent change twice.

I'll solve that bug in a subsequent commit, but let's just test for
the existing behavior for now.

I think the existence of this bug makes the point that we should be
using test_cmp whenever possible, instead of grepping substrings of
expected output. I'm pretty sure we'd have fixed it at the time if we
did that consistently, i.e. "why do we need the same warning twice?"
would have come up during review.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>